add e2e for contract_page

+++ e2e/src/test/java/org/example/gateway/PermissionSetGateway.java
@@ -37,6 +37,26 @@ public class PermissionSetGateway implements Configuration {
 
     }
 
+    public void assignToContract() {
+        String userId = getStandardUserUserId();
+        String permissionSetId = getPermissionSetId(getNamespacePrefix(), PermissionSet.Contract.getName())
+                .orElseThrow(() -> new RuntimeException(String.format("not found permissionSet. namespacePrefix=%s , name=%s", getNamespacePrefix(), PermissionSet.Contract.getName())));
+
+        Result<FindRecordsResult, ErrorsResult> findResult = this.repository.findPermissionSetAssignment(userId, permissionSetId);
+        if (findResult instanceof Result.Failure) {
+            throw new RuntimeException(((Result.Failure) findResult).value().toString());
+        }
+
+        Result.Success<FindRecordsResult> successResult = (Result.Success<FindRecordsResult>) findResult;
+        if (successResult.value().totalSize() == 0) {
+            Result<InsertRecordResult, ErrorsResult> insertResult = this.repository.insertPermissionSetAssignment(new PermissionSetAssignment(userId, permissionSetId));
+            if (insertResult instanceof Result.Failure) {
+                throw new RuntimeException(((Result.Failure) insertResult).value().toString());
+            }
+        }
+
+    }
+
     public void unAssignFromDemo() {
         String userId = getStandardUserUserId();
         String permissionSetId = getPermissionSetId(getNamespacePrefix(), PermissionSet.Demo.getName())
@@ -52,6 +72,30 @@ public class PermissionSetGateway implements Configuration {
             List<String> recordIds = successResult.value().records().stream()
                 .map(record -> record.id())
                 .collect(Collectors.toList());
+
+            Result<DeleteRecordsResult, ErrorsResult> deleteResult = this.repository.deleteRecords(recordIds);
+            if (deleteResult instanceof Result.Failure) {
+                throw new RuntimeException(((Result.Failure) deleteResult).value().toString());
+            }
+        }
+
+    }
+
+    public void unAssignFromContract() {
+        String userId = getStandardUserUserId();
+        String permissionSetId = getPermissionSetId(getNamespacePrefix(), PermissionSet.Contract.getName())
+                .orElseThrow(() -> new RuntimeException(String.format("not found permissionSet. namespacePrefix=%s , name=%s", getNamespacePrefix(), PermissionSet.Demo.getName())));
+
+        Result<FindRecordsResult, ErrorsResult> findResult = this.repository.findPermissionSetAssignment(userId, permissionSetId);
+        if (findResult instanceof Result.Failure) {
+            throw new RuntimeException(((Result.Failure) findResult).value().toString());
+        }
+
+        Result.Success<FindRecordsResult> successResult = (Result.Success<FindRecordsResult>) findResult;
+        if (successResult.value().totalSize() > 0) {
+            List<String> recordIds = successResult.value().records().stream()
+                    .map(record -> record.id())
+                    .collect(Collectors.toList());
 
             Result<DeleteRecordsResult, ErrorsResult> deleteResult = this.repository.deleteRecords(recordIds);
             if (deleteResult instanceof Result.Failure) {
